requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Automat==22.10.0
boto3==1.26.3
botocore==1.29.3
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.3
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.50.0
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
Markdown==3.4.1
msgpack==1.0.4
oauthlib==3.2.2
packaging==21.3
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.6
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
zope.interface==5.5.1
